API_pipeline/requirements.txt
opencv-python
pytesseract
Pillow
pdf2image
numpy
spacy-transformers
paddlepaddle
paddleocr
flask
spacy
spacy-transformers
spaczz
pyarabic
